Support QR code verifications
