+++ docs/iamb.1
@@ -84,6 +84,8 @@ Import and decrypt keys from
 View a list of ongoing E2EE verifications.
 .It Sy ":verify accept [key]"
 Accept a verification request.
+.It Sy ":verify emoji [key]"
+Transition a request to use interactive emoji verification.
 .It Sy ":verify cancel [key]"
 Cancel an in-progress verification.
 .It Sy ":verify confirm [key]"
