+++ src/commands.rs
@@ -35,8 +35,8 @@ use crate::base::{
     RoomField,
     SendAction,
     SpaceAction,
-    VerifyAction,
 };
+use crate::verifications::VerifyAction;
 
 type ProgContext = CommandContext;
 type ProgResult = CommandResult<ProgramCommand>;
@@ -211,6 +211,7 @@ fn iamb_verify(desc: CommandDescription, ctx: &mut ProgContext) -> ProgResult {
                 "cancel" => VerifyAction::Cancel,
                 "confirm" => VerifyAction::Confirm,
                 "mismatch" => VerifyAction::Mismatch,
+                "emoji" => VerifyAction::Emoji,
                 "request" => {
                     let iact = IambAction::VerifyRequest(args.remove(1));
                     let step = CommandStep::Continue(iact.into(), ctx.context.clone());
