+++ src/completions.rs
@@ -373,7 +373,14 @@ fn complete_iamb_keys(
 
 /// Tab completion for `:verify`
 fn complete_iamb_verify(args: Vec<String>, store: &ChatStore) -> Vec<String> {
-    let subcmds = ["request", "accept", "confirm", "cancel", "missmatch"];
+    let subcmds = [
+        "request",
+        "accept",
+        "confirm",
+        "cancel",
+        "missmatch",
+        "emoji",
+    ];
     match args.len() {
         1 => complete_choices(&args[0], &subcmds),
         2 if args[0] == "request" => complete_users(&args[1], store),
