+++ src/main.rs
@@ -28,10 +28,8 @@ use std::sync::atomic::{AtomicUsize, Ordering};
 use std::time::{Duration, Instant};
 
 use clap::{CommandFactory, Parser};
-use matrix_sdk::encryption::verification::{Verification, VerificationRequestState};
 use matrix_sdk::ruma::UserId;
 use matrix_sdk::ruma::api::error::ErrorKind;
-use matrix_sdk::ruma::events::key::verification::VerificationMethod;
 use matrix_sdk::ruma::profile::{ProfileFieldName, ProfileFieldValue};
 use matrix_sdk_crypto::encrypt_room_key_export;
 use modalkit::keybindings::InputBindings;
@@ -89,8 +87,8 @@ mod worker;
 
 #[cfg(test)]
 mod tests;
+mod verifications;
 
-use crate::base::VerifyAction;
 use crate::{
     base::{
         AsyncProgramStore,
@@ -625,97 +623,14 @@ impl Application {
             },
 
             IambAction::Verify(act, flow_id) => {
-                let Some(request) = store.application.verifications.get(&flow_id) else {
-                    return Err(IambError::InvalidVerificationId(flow_id).into());
-                };
-
-                match act {
-                    VerifyAction::Accept => {
-                        request
-                            .accept_with_methods(vec![VerificationMethod::SasV1])
-                            .await
-                            .map_err(IambError::from)?;
-
-                        // Directly start verification since we only support one verification
-                        // method.
-                        if request
-                            .their_supported_methods()
-                            .is_some_and(|methods| methods.contains(&VerificationMethod::SasV1))
-                        {
-                            request.start_sas().await.map_err(IambError::from)?;
-                        }
-
-                        Some(InfoMessage::from("Accepted verification request"))
-                    },
-                    VerifyAction::Cancel => {
-                        request.cancel().await.map_err(IambError::from)?;
-                        Some(InfoMessage::from("Cancelled verification"))
-                    },
-                    VerifyAction::Confirm => {
-                        let VerificationRequestState::Transitioned {
-                            verification: Verification::SasV1(sas),
-                        } = request.state()
-                        else {
-                            let msg = "Can only confirm in-progress verifications!";
-                            let err = UIError::Failure(msg.into());
-                            return Err(err);
-                        };
-
-                        if !sas.can_be_presented() {
-                            let msg = "Can only confirm in-progress verifications!";
-                            let err = UIError::Failure(msg.into());
-                            return Err(err);
-                        }
-
-                        sas.confirm().await.map_err(IambError::from)?;
-
-                        Some(InfoMessage::from("Confirmed verification"))
-                    },
-                    VerifyAction::Mismatch => {
-                        let VerificationRequestState::Transitioned {
-                            verification: Verification::SasV1(sas),
-                        } = request.state()
-                        else {
-                            let msg = "Can only reject in-progress verifications!";
-                            let err = UIError::Failure(msg.into());
-                            return Err(err);
-                        };
-
-                        if !sas.can_be_presented() {
-                            let msg = "Can only reject in-progress verifications!";
-                            let err = UIError::Failure(msg.into());
-                            return Err(err);
-                        }
-
-                        sas.mismatch().await.map_err(IambError::from)?;
-
-                        Some(InfoMessage::from("Rejected verification"))
-                    },
-                }
+                return verifications::iamb_verify(act, flow_id, store).await;
             },
             IambAction::VerifyRequest(user_id) => {
                 let Ok(user_id) = <&UserId>::try_from(user_id.as_str()) else {
                     return Err(IambError::InvalidUserId(user_id).into());
                 };
-                let enc = self.worker.client.encryption();
 
-                let Some(identity) =
-                    enc.get_user_identity(user_id).await.map_err(IambError::from)?
-                else {
-                    let msg = format!("Could not find identity information for {user_id}");
-                    let err = UIError::Failure(msg);
-                    return Err(err);
-                };
-
-                let methods = vec![VerificationMethod::SasV1];
-                let request = identity.request_verification_with_methods(methods);
-                let request = request.await.map_err(IambError::from)?;
-
-                let flow_id = request.flow_id().to_owned();
-                store.application.verifications.insert(flow_id, request);
-
-                let info = format!("Sent verification request to {user_id}");
-                Some(InfoMessage::from(info))
+                return verifications::iamb_verify_request(user_id, store).await;
             },
         };
